# swagger.json
{"openapi":"3.0.0","info":{"title":"My API","version":"1.0.0","description":"API documentation"},"servers":[{"url":"https://social-api-cvd9b5dmbmacfrcu.francecentral-01.azurewebsites.net/"}],"paths":{"/media/upload":{"post":{"summary":"Upload a media file","description":"Upload a media file (image, video, etc.) to Azure Blob Storage.","tags":["Media"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary","description":"The file to upload."}}}}}},"responses":{"201":{"description":"File uploaded successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"},"fileUrl":{"type":"string","description":"The URL of the uploaded file."}}}}}},"400":{"description":"File is required."},"500":{"description":"Failed to upload file."}}}},"/media/download/{fileName}":{"get":{"summary":"Download a media file","description":"Download a file from Azure Blob Storage.","tags":["Media"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"fileName","required":true,"schema":{"type":"string"},"description":"The name of the file to download."}],"responses":{"200":{"description":"File downloaded successfully.","content":{"application/octet-stream":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"File not found."},"500":{"description":"Failed to download file."}}}},"/post":{"post":{"summary":"Create a new post","description":"Create a new post associated with the authenticated user.","tags":["Posts"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"content":{"type":"string","description":"The content of the post."}}}}}},"responses":{"201":{"description":"Post created successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userId":{"type":"integer"}}}}}},"400":{"description":"Content is required."},"500":{"description":"Failed to create post."}}}},"/post/{id}":{"get":{"summary":"Get a specific post","description":"Retrieve a post by its ID.","tags":["Posts"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"The ID of the post."}],"responses":{"200":{"description":"The requested post.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userId":{"type":"integer"}}}}}},"404":{"description":"Post not found."},"500":{"description":"Failed to fetch post."}}}},"/user":{"post":{"summary":"Create a new user","description":"Create a new user in the system.","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"isPrivate":{"type":"boolean"},"isAdmin":{"type":"boolean"}}}}}},"responses":{"201":{"description":"User created successfully.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"isPrivate":{"type":"boolean"},"isAdmin":{"type":"boolean"}}}}}},"400":{"description":"Username and password are required."},"500":{"description":"Failed to create user."}}},"get":{"summary":"Get all users","description":"Retrieve a list of all users. Protected route.","tags":["Users"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"A list of users.","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"isPrivate":{"type":"boolean"},"isAdmin":{"type":"boolean"}}}}}}},"500":{"description":"Failed to fetch users."}}}},"/user/login":{"post":{"summary":"User login","description":"Authenticate a user and return a JWT token.","tags":["Users"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Login successful.","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"},"user":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"isPrivate":{"type":"boolean"},"isAdmin":{"type":"boolean"}}}}}}}},"400":{"description":"Username and password are required."},"401":{"description":"Invalid username or password."}}}},"/user/{id}":{"get":{"summary":"Get a specific user","description":"Retrieve a user by their ID. Protected route.","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"The user ID."}],"responses":{"200":{"description":"The requested user.","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"isPrivate":{"type":"boolean"},"isAdmin":{"type":"boolean"}}}}}},"404":{"description":"User not found."},"500":{"description":"Failed to fetch user."}}},"put":{"summary":"Update a user","description":"Update a user's details by their ID. Protected route.","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"The user ID."}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"200":{"description":"User updated successfully."},"400":{"description":"No updates provided."},"500":{"description":"Failed to update user."}}},"delete":{"summary":"Delete a user","description":"Delete a user by their ID. Protected route.","tags":["Users"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"},"description":"The user ID."}],"responses":{"204":{"description":"User deleted successfully."},"500":{"description":"Failed to delete user."}}}}},"components":{},"tags":[]}